local language shared library and tests
