+++ packages/shared/src/lib/geolocation.ts
@@ -545,37 +545,6 @@ const extendedCurrencySymbol = {
     International: '$',
 };
 
-type LocalLanguageHeader = {
-    header: string;
-};
-
-type LocalLanguageHeaders = Record<string, LocalLanguageHeader>;
-
-const localLanguagesHeaders: LocalLanguageHeaders = {
-    FR: {
-        header: 'Faites partie de notre aventure européenne !',
-    },
-    DE: {
-        header: 'hallo (German)',
-    },
-    IT: {
-        header: 'Prendi parte alla nostra nuova avventura europea!',
-    },
-    NL: {
-        header: 'hallo (Dutch)',
-    },
-    SE: {
-        header: 'halla',
-    },
-    SP: {
-        header: 'hola',
-    },
-};
-
-export const countryCodeToLocalLanguageHeader = (countryCode?: string): string => {
-    return localLanguagesHeaders[countryCode].header || '';
-};
-
 export const countryCodeToCountryGroupId = (countryCode?: string): CountryGroupId => {
     const availableCountryGroupIds = Object.keys(countryGroups) as CountryGroupId[];
